refactor: restructure ProjectCard component and tests
- Moved ProjectCard component to its own folder for better organisation.
- Updated imports in Projects.tsx and ProjectCard test file accordingly.
- Modified test for technology icons to use `getByTitle` for improved accuracy.
- Adjusted mock project data to update the technology icon structure.
